@@ -9,12 +9,12 @@ Usage:
 """
 
 import argparse
+from pathlib import Path
 import subprocess  # nosec B404 # Safe use with hardcoded commands
 import sys
-from pathlib import Path
 
 
-def run_coverage(full_report=False, check_only=False, threshold=80):
+def run_coverage(full_report: bool = False, check_only: bool = False, threshold: int = 80) -> bool:
     """Run pytest with coverage and provide smart output."""
 
     if check_only:
@@ -55,9 +55,9 @@ def run_coverage(full_report=False, check_only=False, threshold=80):
         ]
 
     try:
-        result = subprocess.run(  # nosec B603 # Safe use with hardcoded command list
+        result = subprocess.run(
             cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent
-        )
+        )  # nosec B603 # Safe use with hardcoded command list
 
         if check_only:
             # Extract just the coverage percentage
@@ -69,17 +69,11 @@ def run_coverage(full_report=False, check_only=False, threshold=80):
                         if "%" in part:
                             coverage = int(part.replace("%", ""))
                             if coverage < threshold:
-                                print(
-                                    f"❌ Coverage: {coverage}% (below {threshold}% threshold)"
-                                )
-                                print(
-                                    "💡 Run 'python scripts/coverage.py --full' to see what needs testing"
-                                )
+                                print(f"❌ Coverage: {coverage}% (below {threshold}% threshold)")
+                                print("💡 Run 'python scripts/coverage.py --full' to see what needs testing")
                                 return False
                             else:
-                                print(
-                                    f"✅ Coverage: {coverage}% (above {threshold}% threshold)"
-                                )
+                                print(f"✅ Coverage: {coverage}% (above {threshold}% threshold)")
                                 return True
         else:
             # Show the output
@@ -95,13 +89,9 @@ def run_coverage(full_report=False, check_only=False, threshold=80):
                         if "%" in part:
                             coverage = int(part.replace("%", ""))
                             if coverage < threshold:
-                                print(
-                                    f"\n💡 Coverage is {coverage}% (below {threshold}%)"
-                                )
+                                print(f"\n💡 Coverage is {coverage}% (below {threshold}%)")
                                 if not full_report:
-                                    print(
-                                        "   Run 'python scripts/coverage.py --full' for detailed missing lines"
-                                    )
+                                    print("   Run 'python scripts/coverage.py --full' for detailed missing lines")
                             break
 
         return result.returncode == 0
@@ -111,26 +101,15 @@ def run_coverage(full_report=False, check_only=False, threshold=80):
         return False
 
 
-def main():
+def main() -> None:
     parser = argparse.ArgumentParser(description="Run coverage analysis")
-    parser.add_argument(
-        "--full", action="store_true", help="Show full detailed coverage report"
-    )
-    parser.add_argument(
-        "--check", action="store_true", help="Just check if coverage meets threshold"
-    )
-    parser.add_argument(
-        "--threshold",
-        type=int,
-        default=80,
-        help="Coverage threshold percentage (default: 80)",
-    )
+    parser.add_argument("--full", action="store_true", help="Show full detailed coverage report")
+    parser.add_argument("--check", action="store_true", help="Just check if coverage meets threshold")
+    parser.add_argument("--threshold", type=int, default=80, help="Coverage threshold percentage (default: 80)")
 
     args = parser.parse_args()
 
-    success = run_coverage(
-        full_report=args.full, check_only=args.check, threshold=args.threshold
-    )
+    success = run_coverage(full_report=args.full, check_only=args.check, threshold=args.threshold)
 
     sys.exit(0 if success else 1)
 
