@@ -1,4 +1,5 @@
-""" A client library for accessing Span """
+"""A client library for accessing Span"""
+
 from .client import AuthenticatedClient, Client
 
 __all__ = (
