@@ -1,1 +1,1 @@
-""" Contains endpoint functions for accessing the API """
+"""Contains endpoint functions for accessing the API"""
