@@ -22,9 +22,8 @@ def _get_kwargs(
         "url": "/api/v1/wifi/connect",
     }
 
-    _body = body.to_dict()
+    _kwargs["json"] = body.to_dict()
 
-    _kwargs["json"] = _body
     headers["Content-Type"] = "application/json"
 
     _kwargs["headers"] = headers
